Move receive_task function to rxpipe superclass
