+++ src/rx_pipe/__init__.py
@@ -1,3 +1,4 @@
+import asyncio
 import logging
 
 from ..task_executor import TaskExecutor
@@ -17,6 +18,19 @@ class RXPipe:
         self.listen_address = listen_address
         self.listen_port = listen_port
         self.task_executor = task_executor
+
+    async def receive_task(self, reader : asyncio.StreamReader, writer : asyncio.StreamWriter) -> None:
+        """Generic callback function which passes offloaded task directly to the task executor.
+
+        """
+        input_data = await reader.read()
+
+        result_data = await self.task_executor.execute_task(input_data)
+
+        writer.write(result_data)
+        await writer.drain()
+        writer.close()
+        self.logger.info("Processed task")
 
     def set_logger(self, logger : logging.Logger):
         self.logger = logger
